moved registration and login views, forms, templates outside of reporting module
