+++ app/templates/emails/registration.html
