+++ app/templates/login.html
@@ -1,4 +1,4 @@
-{% extends "reporting/base.html" %}
+{% extends "base.html" %}
 {% block sub_heading %}Login{% endblock %}
 {% block content %}
 {% from "forms/macros.html" import render_field %}
