+++ app/templates/reporting/base.html
@@ -1,17 +1,3 @@
 {% extends "base.html" %}
 {% block title %}Incidental Reporting{% endblock %}
-{% block account %}
-{% if g.user %}
-<div class="btn-group pull-right">
-    <a class="btn btn-primary" href="#"><i class="icon-user"></i> {{ g.user.name }}</a>
-    <a class="btn btn-primary dropdown-toggle" data-toggle="dropdown" href="#"><span class="icon-caret-down"></span></a>
-    <ul class="dropdown-menu">
-        <!--<li><a href="/profile"><i class="icon-pencil"></i> Account</a></li>
-                        <li><a href="#"><i class="icon-lock"></i> Privacy</a></li>
-                        <li class="divider"></li>-->
-        <li><a href="/reporting/logout"><i class="icon-signout"></i> Logout</a></li>
-    </ul>
-</div>
-{% endif %}
-{% endblock %}
 {% block heading %}Incidental Reporting{% endblock %}
