--- app/templates/reporting/confirm_rename.html
@@ -1,13 +1,0 @@
-{% extends "reporting/base.html" %}
-{% from "forms/macros.html" import render_field %}
-{% block sub_heading %}Clinically Relevant Scan Types{% endblock %}
-{% block content %}
-<h5>Showing {{ num_showing[0] }} of {{ num_showing[1] }} unconfirmed scan types</h5>
-<form method="POST" action=" {{ url_for('reporting.confirm_scan_types') }}" class="form">
-    {{ form.csrf_token }}
-    {{ render_field(form.clinical_scans, class="form-checkbox") }}
-    {{ form.viewed_scan_types() }}
-    <input type="submit" value="Confirm" class="button green">
-    &nbsp;&nbsp;<a href="{{ url_for('reporting.index') }}">Cancel</a>
-</form>
-{% endblock %}
