+++ mbi-flask.service
@@ -1,0 +1,17 @@
+# sylink to /etc/systemd/system/mbi-flask.service on Debian/Ubuntu systems
+
+[Unit]
+Description=MBI Flask
+Requires=docker.service
+After=docker.service
+
+[Service]
+Type=oneshot
+RemainAfterExit=yes
+WorkingDirectory=/home/flask/mbi-flask
+ExecStart=/usr/local/bin/docker-compose up -d
+ExecStop=/usr/local/bin/docker-compose down
+TimeoutStartSec=0
+
+[Install]
+WantedBy=multi-user.target
