+++ web/config.template
@@ -8,17 +8,17 @@ DEBUG = True
 SQLALCHEMY_TRACK_MODIFICATIONS = False
 TEST = False
 
-SQLALCHEMY_DATABASE_URI = 'sqlite:////database.db'
+SQLALCHEMY_DATABASE_URI = 'sqlite:////databases/app.db'
 
 SECRET_KEY = '$FLASK_SECRET_KEY'
 
 SOURCE_XNAT_URL = 'https://mbi-xnat.erc.monash.edu.au'
-SOURCE_XNAT_USER = 'mbi_flask'
-SOURCE_XNAT_PASSWORD = $SOURCE_XNAT_PASSWORD
+SOURCE_XNAT_USER = '$SOURCE_XNAT_USER'
+SOURCE_XNAT_PASSWORD = '$SOURCE_XNAT_PASSWORD'
 
-TARGET_XNAT_URL = 'https://172.22.2.13'
-TARGET_XNAT_USER = 'tclose'
-TARGET_XNAT_PASSWORD = $TARGET_XNAT_PASSWORD
+TARGET_XNAT_URL = 'http://172.22.2.13'
+TARGET_XNAT_USER = '$TARGET_XNAT_USER'
+TARGET_XNAT_PASSWORD = '$TARGET_XNAT_PASSWORD'
 
 
 TARGET_XNAT_PROJECT = 'MBIReporting'
@@ -30,7 +30,7 @@ WTF_CSRF_SECRET_KEY = '$WTF_CSRF_SECRET_KEY'
 
 ADMIN_EMAIL = 'mbi-informatics@monash.edu'
 
-FILEMAKER_IMPORT_FILE = '/filemaker-export.csv'
+FILEMAKER_IMPORT_FILE = '/to-import/filemaker-export.csv'
 
 AUTO_LOGOUT_PERIOD = timedelta(minutes=30)
 
@@ -42,9 +42,9 @@ REPORT_INTERVAL = 365
 UPLOADED_SIGNATURE_DEST = '/uploads'
 
 
-MAIL_SERVER = 'smtp.gmail.com'
+MAIL_SERVER = smtp.gmail.com
 MAIL_PORT = 465
 MAIL_USE_SSL = True
-MAIL_USERNAME = 'tom.close@monash.edu'
+MAIL_USERNAME = '$MAIL_USER'
 MAIL_PASSWORD = '$MAIL_PASSWORD'
 MAIL_DEFAULT_SENDER = ADMIN_EMAIL
